requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
inflection==0.5.1
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
orjson==3.9.15
packaging==24.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
piccolo==1.4.2
piccolo-api==1.1.0
piccolo_admin==1.3.1
platformdirs==4.2.0
priority==2.0.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
targ==0.3.8
taskgroup==0.0.0a4
tomli==2.0.1
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.28.0
uvloop==0.19.0
wcwidth==0.2.13
wsproto==1.2.0
